fix: moved forms to separate files, typescript fixes
