+++ src/components/frame/v5/pages/UserAdvancedPage/partials/FeesForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './FeesForm';
