+++ src/components/frame/v5/pages/UserAdvancedPage/partials/RpcForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RpcForm';
