--- src/components/frame/v5/pages/UserAdvancedPage/partials/UserAdvancedSettings/types.ts
@@ -1,5 +1,0 @@
-export interface AdvancedSettingsFields {
-  metatransactions: boolean;
-  decentralizedModeEnabled: boolean;
-  customRpc: string;
-}
