--- src/components/frame/v5/pages/UserAdvancedPage/types.ts
@@ -1,5 +1,0 @@
-export interface AdvancedSettingsFields {
-  metatransactions: boolean;
-  decentralizedMode: boolean;
-  customRPC: string;
-}
